Included sample pac man project and example maps.
Checkout the Readme in the maps folder before creating your own map.
Look to the Program 5 (as plain text) text file in the Program 5 folder
to view the original code outside of VS. The project relies on an
Allegro 4 library to function, so I am unsure how helpful it will be.
This project should only be used as a rough reference.